newstyle.css
Added flexbox to the website by using justify-content and nowrap. No wrap allows all flex items to try and fit on one line at the bottom. I was able to center all of the sections of the html headers using justify-content:center. Added the clarify fitness logo as apart of the website also.