Add files via upload
Main.css and html is the newly improved code for the website's landing page. The login is for the login page of the website used to register in the clients.mindbodyonline.com subdomain. The class schedule is the new and improved class schedule format located in the  clients.mindbodyonline.com subdomain also. A/B Testing will also be added, but only to the original Clarity Fitness website, not this new modified version.